Switch to separate development settings file
Specifying CONN_MAX_AGE in development causes database connections to
pile up because the development server uses a new connection for each
request. However, a positive value for CONN_MAX_AGE is desirable in
production because it reduces latency for most requests. Therefore, we
need separate Django settings for production and development in order to
provide different values for CONN_MAX_AGE.