Ensure that development version of Ashlar is used on restart
A version of Ashlar is installed during the Docker image building
process, but that version comes from the `develop` branch on GitHub. In
order to support local development, Ashlar needs to be installed in
"editable" mode. From there `--src` is used to determine where the
contents of the "editable" payload will live.

In this scenario, `--src` is set to `/opt`, which is where the source
directory is mounted in the container.