Create separate service for fetching boundary info.
Fixes #146.
The app has been fetching the GeoJSON for the currently selected user boundary before loading the rest of the app,
causing long load times. The bounds themselves are not needed in the UI, and will only be used in filtering
for some reports. This fetches the boundary metadata instead, which is sufficient for the dropdown.