Zoom leaflet maps to bounds selection
Creates a pluggable directive that can be attached to an instance of a
leaflet-map directive. Its only purpose is to listen for boundary
changes and zoom the map to focus on the new boundary.

Additionally, stops loading full GeoJSON for every boundary and loads
just the bounding box instead; this provides a significant boost in page
load times.