Make record type dropdown visibility configurable
A web config setting of `recordType.visible` controls visibility.
If the record type is configured to not be visible, the record type
is chosen by using the config setting of `recordType.primaryLabel`.

A seting for `recordType.secondaryLabel` has also been added, though
it is not yet in use. It will be used when displaying interventions.