Add service for tracking initial state
All pages of the app rely on the record type, geography, and boundary.
This adds a central place for keeping track of that initialization.
Previously, all components were individually requesting data on initial
page load, which led to many duplicate requests and race conditions.
This eliminates the duplicate requests and fixes a couple bugs.

Fixes #198, #228