Add audit logging for changes to records
Audit logs are generated on each create, update, or delete operation.
The audit-log endpoint is accessible only by admin users, and even for
them it is read-only. The endpoint is unpaginated by default, and
supports filtering by date ranges. Supports CSV output, since the
desired functionality is for users to be able to download monthly CSVs
of audit log activity.