Add generate black spots input script
This script reads in a shapefile of roads and a csv of records,
separates out intersections and divides up segements, matches
up records to nearest segments, aggregates counts by year, and
writes the data to a shapefile and a csv file. The output is
intended to be used in a machine learning algorithm to generate
black spots.