Refactor doFilter for more control and usability
QueryBuilder.djangoQuery had a 'doFilter' toggle to control whether to
apply filters or not. This splits it into doAttrFilters and doJsonFilters
to enable filtering on record attributes (which all records share)
separately from the json filters (which are RecordSchema-specific).

It also moves them to the end of the argument list so that they can be
optional (defaulting to true) since we almost always want filtering.