Clean up record type nomenclature
The primary record was being called lots of things, now it will mostly be
called 'primary'.

Also, in all other controllers ctl.recordType is the object, but in this
one it was the uuid, which was both confusing and inconvenient (since it
meant we had to store the label and plural label in separate variables).
So this changes it to be the object.