Fix record details name/key error
The fields in record list view had started breaking because PR #405 took
ctl.detailsPropertyName to be a title when in fact it's a key. This
changes it to ctl.detailsPropertyKey.
This also rips out ctl.detailsProperty there and in two other places
where it was being set with the title of the details property but
not used except in error.