Add interventions tool tab + consolidate tool tabs
Adds a tool tab to the map to enable adding and exporting interventions.
The files are called 'interventions' but the labels shown to the user
are taken from the secondary RecordType.

Also consolidates the tool tabs (this, graphs, and export) into a
'tools' directory, since they're quite similar and somewhat connected.