Make options promises coordinate properly
The two different queries used to get aggregation options are
independent, but the returned promise shouldn't resolve until everything
is guaranteed to be done, which it wasn't the case because the
intermediate functions weren't returning anything.  I believe this setup
will only resolve when fully done but impose no unnecessary blocking
internally.