Move boundary filter into QueryBuilder
Several different components have been checking BoundaryState and
manually adding a boundary filter to record queries when one is set.
This adds a doBoundaryFilter toggle to the QueryBuilder methods, and
modifies the affected pieces to use that rather than do it themselves.

Since that brings us to three separately-controllable filter categories,
I changed QueryBuilder.djangoQuery() and QueryBuilder.assembleParams()
to take a filterConfig object rather then the long series of booleans
that are annoying inside the service and nigh-impossible to keep
straight when invoking it.

Custom reports are the one thing I'm aware of that still isn't boundary-
aware, since the interaction between the selected boundary and
aggregating by geography requires a little bit of figuring out and would
need to be communicated to the user on the custom report modal.