Make navbar respond to authentication changes
Parts of the navbar weren't switching to reflect that a user had logged
in. This consolidates the user-related variables and updates them when
someone logs in. (Turns out there are at least a few, maybe several other
controllers with the same problem, but at least this one is better now.)