Fix race condition with translations
The translations performed in javascript with `$translate.instant`
need to ensure the translation file is ready at the time of translation.
`InitialState` is an object that many controllers rely on in order
to wait for all basic resources to be available before initialization.
`InitialState` has been modified to also check for translations being
ready. This covers the majority of potential translation race conditions.

There are a couple modules that aren't using `InitialState` and also
rely on `$translate.instant`. In these cases, `$translate.onReady` is
being used directly. It's hard to test these locally, since translation
files are loaded so quickly, the race condition is not apparent. I'll
test this once it makes its way to staging and will fix any additional
sections that show themselves, if there are any.