Ensure batch cron jobs are not run multiple times
This commit adds an initial heimdall configuration. Heimdall is a program
that wraps an executable program inside of an exclusive lock provided by
PostgreSQL's `pg_try_advisory_lock`. This results in the ability to wrap
cron jobs in heimdall across multiple worker machines and have only one
instance of the job run.