Optionally display black spots dashboard widget
When black spots visibility is disabled, the black spots
widget is not displayed, and the other widgets are shifted
around accordingly. Also some minor changes were made to
wording and font size for more consistency, and a greater
likelihood that the labels aren't cut off on small screen
sizes (tested with 1024x768).