Fix problems with black spots calculation task
These fixes allow the task to run to completion with a smaller
OSM input. Optimizations for a larger input will be done in a
separate PR. This was tested by setting `osm_extract_url` to:
'http://osm-extracted-metros.s3.amazonaws.com/manila.osm.pbf',
which uses the Manila boundary (rather than the full Philippines).