Add BlackSpotConfig model and API endpoints
The API can only be used to update the config, not create or delete
configs, because the BlackSpotConfig is meant to be treated as a
singleton.

A data migration is included to generate a default BlackSpotConfig,
although the API list endpoint will also generate one if it doesn't
already exist.