Fix directory used for writing black spot output
The shared directory being used to communicate the R output back
to the celery task (/opt/app) was only valid while in development.
In order to work in staging/production this has been switched to
the media directory, which is available in both development and
staging/production, and is actually where all the other batch
output files were already being written.