Fix color alignment in incident record detail
The right-to-left class successfully flips all the text on the page,
but SVG coordinates are still in left-to-right, so vehicle color
rectangles wound up left-aligned in their enclosing SVGs, which was
way out of place.

This fix conditions a shift to the rectangles' x attributes
on right-to-left-ness and the size of the rectangle to
appropriately right-align the SVG elements.