Add a number type to schema-builder
Validation on the Android app depends on a better number type
than text formatted as a number. This commit adds that number
type and and renderers for displaying that information in
detail views.

Types in `incident_schema_v3.json` were changed to reflect
the existence of this new field type.

Additionally, one test was changed so that objects containing
the same keys and values pass even if the order of those keys
and values gets scrambled somehow.

Fixes #562.