Include rendering for numbers and integers
In #598, I didn't include the directives and controllers for
rendering number and integer fields in the deployment index
file. Naturally, they didn't render, even though with my
non-deployment index.html they did. This fixes that and breaks
out the controllers/directives for number and integer types
in case we want to do things differently with them in the
future.