Add the email / username of the last user to edit
Adds the email / username of the last user to edit a record to the
detail view for authenticated users. Because users can be deleted, this
falls back to the username if the user has been deleted, which is the
functionality provided by the RecordAuditLogEntry class. In cases where
the user was generated using Google authentication, the username is
unlikely to be helpful (e.g. 116570040541971129286) but it will at least
indicate that the user no longer exists.