Support tagged releases in Travis, Ansible
This configures Travis to deploy Docker containers with tags matching
the built tag on the master branch. It also adds an Ansible variable
called `app_version` that can be used to specify which Docker container
gets deployed when running production deployments. In development, tags
are not used, only `latest` is used.