Restart driver-app service only after we have the new image
This changes provisioning so that instead of restarting the service
after all the tasks have run via the handler, we now do so explicitly in
a dedicated task _after_ we pull/build the new image and _before_ we use
the app container.

Previously, we were running into issues where we would run migrations
against the previous version of the container.