Fix bug with non-ASCII characters in filter name
The `jsonb` query parameter containing the filter name (with non-ASCII
characters in it) comes into the system as type 'str'. Previously, the
SQL produced by `generate_query_sql` (also a string) was then re-encoded
as a string, causing an error for non-ASCII characters.

This fixes the bug by having `generate_query_sql` return utf-8, and then
dealing with that from there out (re-encoding as a string before saving
to Redis).