Fix relationship value not showing on record modal
The fix here is to remove the one-time binding syntax for `record`, as
is also performed above in the `driver-details-single` section. It appears
there is a sort of race condition, where the record data can possibly be
altered after its initial load. I imagine that's why this same change was
made to `driver-details-single`.

With the one-time binding syntax in place, there was some non-determinism,
where sometimes the relationship data would be loaded, but other times it
wouldn't be (because the data wasn't present on the record yet).